backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
feedparser==6.0.11
beautifulsoup4==4.12.3
httpx==0.26.0
python-dateutil==2.8.2
google-generativeai==0.3.2
pydantic==2.5.3
lxml==5.1.0
